requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
pydub>=0.25.1
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio>=3.1.1
transformers>=4.35.0
speechbrain>=0.5.15
tensorflow>=2.14.0  # for YAMNet
tensorflow-hub>=0.14.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.1
soundfile>=0.12.1 